+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Types of changes:
 ## Next
 
 ### Added
-
+- `<PageHeader/>` Testkit - expose puppeteer testkit [#3536](https://github.com/wix/wix-style-react/pull/3536)
 - `<ListItemAction/>` - add classname prop for extensibility [#3543](https://github.com/wix/wix-style-react/pull/3543)
 
 ### Docs